% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_identification.R
\name{are_interacting}
\alias{are_interacting}
\title{Check if the proteins encoded by two genes interact}
\usage{
are_interacting(gene1 = NULL, gene2 = NULL, edgelist_ppi = NULL)
}
\arguments{
\item{gene1}{Character of gene ID for gene 1.}

\item{gene2}{Character of gene ID for gene 2.}

\item{edgelist_ppi}{A 2-column data frame with IDs of genes that encode
each protein in the interacting pair.}
}
\value{
Logical indicating whether the proteins encoded
by \strong{gene1} and \strong{gene2} interact in a PPI network.
}
\description{
Check if the proteins encoded by two genes interact
}
\examples{
data(gma_ppi)
gene1 <- "Glyma.19G213200"
gene2 <- "Glyma.01G004300"
are_paralogs(gene1, gene2, gma_ppi)
}
