% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_identification.R
\name{are_paralogs}
\alias{are_paralogs}
\title{Check if two genes are paralogs}
\usage{
are_paralogs(gene1 = NULL, gene2 = NULL, paralogs = NULL)
}
\arguments{
\item{gene1}{Character of gene ID for gene 1.}

\item{gene2}{Character of gene ID for gene 2.}

\item{paralogs}{A 2-column data frame with gene IDs for each paralog
in the paralog pair.}
}
\value{
Logical indicating whether \strong{gene1} and \strong{gene2} are
paralogs.
}
\description{
Check if two genes are paralogs
}
\examples{
data(gma_paralogs)
paralogs <- gma_paralogs[, 1:2]
gene1 <- "Glyma.01G011500"
gene2 <- "Glyma.01G011600"
are_paralogs(gene1, gene2, paralogs)
}
