% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_identification.R
\name{count_lambda}
\alias{count_lambda}
\title{Count the frequency of lambda motifs}
\usage{
count_lambda(edgelist = NULL, paralogs = NULL)
}
\arguments{
\item{edgelist}{A 2-column data frame with regulators in column 1 and
targets in column 2.}

\item{paralogs}{A 2-column data frame with gene IDs for each paralog
in the paralog pair.}
}
\value{
Numeric scalar with the frequency of lambda motifs in the network.
}
\description{
Count the frequency of lambda motifs
}
\examples{
data(gma_grn)
data(gma_paralogs)
edgelist <- gma_grn[1:500, 1:2] # reducing for test purposes
paralogs <- gma_paralogs[gma_paralogs$type == "WGD", 1:2]
count <- count_lambda(edgelist, paralogs)
}
