% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_identification.R
\name{find_delta}
\alias{find_delta}
\title{Find delta motifs}
\usage{
find_delta(edgelist = NULL, paralogs = NULL, edgelist_ppi = NULL)
}
\arguments{
\item{edgelist}{A 2-column data frame with regulators in column 1 and
targets in column 2.}

\item{paralogs}{A 2-column data frame with gene IDs for each paralog
in the paralog pair.}

\item{edgelist_ppi}{A 2-column data frame with IDs of genes that encode
each protein in the interacting pair.}
}
\value{
A list of data frames, each data frame being an egde list with
the interactions in the motif.
}
\description{
Find delta motifs
}
\examples{
data(gma_grn)
data(gma_paralogs)
data(gma_ppi)
edgelist <- gma_grn[500:1000, 1:2] # reducing for test purposes
paralogs <- gma_paralogs[gma_paralogs$type == "WGD", 1:2]
edgelist_ppi <- gma_ppi
motifs <- find_delta(edgelist, paralogs, edgelist_ppi)
}
