% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gma_grn}
\alias{gma_grn}
\title{Sample soybean GRN}
\format{
A 3-column data frame with node1, node2, and edge weight.
}
\usage{
data(gma_grn)
}
\description{
The GRN was inferred with {BioNERO} using expression data from
Libault et al., 2010, and Severin et al., 2010.
}
\examples{
data(gma_grn)
}
\references{
Severin, A. J., Woody, J. L., Bolon, Y. T., Joseph, B., Diers, B. W.,
Farmer, A. D., ... & Shoemaker, R. C. (2010).
RNA-Seq Atlas of Glycine max: a guide to the soybean transcriptome.
BMC plant biology, 10(1), 1-16.

Libault, M., Farmer, A., Joshi, T., Takahashi, K., Langley, R. J.,
Franklin, L. D., ... & Stacey, G. (2010).
An integrated transcriptome atlas of the crop model Glycine max,
and its use in comparative analyses in plants.
The Plant Journal, 63(1), 86-99.
}
\keyword{datasets}
