% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gma_paralogs}
\alias{gma_paralogs}
\title{Soybean (Glycine max) duplicated genes}
\format{
A 3-column data frame with duplicate 1, duplicate 2, and
duplication type
}
\usage{
data(gma_paralogs)
}
\description{
The repertoire of soybean paralogs was retrieved from
Almeida-Silva et al., 2020.
}
\examples{
data(gma_paralogs)
}
\references{
Almeida-Silva, F., Moharana, K. C., Machado, F. B., &
Venancio, T. M. (2020). Exploring the complexity of soybean (Glycine max)
transcriptional regulation using global gene co-expression networks.
Planta, 252(6), 1-12.
}
\keyword{datasets}
