% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gma_ppi}
\alias{gma_ppi}
\title{Sample soybean PPI network}
\format{
A 2-column data frame with node1 and node2.
}
\usage{
data(gma_ppi)
}
\description{
PPI were retrieved from the STRING database and filtered to keep only
medium confidence edges and nodes in the GRN.
}
\examples{
data(gma_ppi)
}
\keyword{datasets}
