% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_identification.R
\name{find_bifan}
\alias{find_bifan}
\title{Find bifan motifs}
\usage{
find_bifan(edgelist = NULL, paralogs = NULL)
}
\arguments{
\item{edgelist}{A 2-column data frame with regulators in column 1 and
targets in column 2.}

\item{paralogs}{A 2-column data frame with gene IDs for each paralog
in the paralog pair.}
}
\value{
A character vector with bifan motifs represented
in the format \strong{regulator1, regulator2->target1, target2}.
}
\description{
Find bifan motifs
}
\examples{
data(gma_grn)
data(gma_paralogs)
edgelist <- gma_grn[600:1400, 1:2] # reducing for test purposes
paralogs <- gma_paralogs[gma_paralogs$type == "WGD", 1:2]
paralogs <- rbind(
paralogs,
data.frame(duplicate1 = "Glyma.01G177200", 
           duplicate2 = "Glyma.08G116700")
)
bifan <- find_bifan(edgelist, paralogs)
}
