% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nulls}
\alias{nulls}
\title{Null distribution of motif frequencies for vignette data set}
\format{
A list of numeric vectors with the motif frequencies in each
simulated network.
}
\usage{
data(nulls)
}
\description{
Data were filtered exactly as demonstrated in the vignette.
}
\examples{
data(nulls)
}
\keyword{datasets}
